requirements.txt
altgraph==0.17.3
APScheduler==3.6.3
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
cachetools==4.2.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.6
Django==3.2
django-debug-toolbar==2.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
Faker==12.0.1
flake8==3.9.2
flake8-docstrings==1.6.0
future==0.18.2
idna==2.8
iniconfig==1.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==9.1.0
packaging==21.3
pefile==2022.5.30
Pillow==9.3.0
pluggy==0.13.1
prettytable==3.5.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.11
PyJWT==2.7.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.26.0
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
six==1.14.0
snowballstemmer==2.2.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
toml==0.10.2
tornado==6.3
tqdm==4.64.1
typing==3.7.4.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.25.11
vboxapi==1.0
wcwidth==0.2.5
